backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pydantic==2.5.1
sqlalchemy==2.1.0
alembic==1.11.1
psycopg2-binary==2.9.7
pytest==7.4.0
httpx==0.24.1
requests==2.31.0
python-dotenv==1.0.0
